incident-management-agents/azure_functions/requirements.txt
# Azure Functions requirements
azure-functions>=1.21.3
azure-identity>=1.19.0

# Azure Management SDKs
azure-mgmt-compute>=33.0.0
azure-mgmt-web>=7.3.1
azure-mgmt-resource>=23.2.0
azure-mgmt-network>=27.0.0
azure-mgmt-storage>=21.3.0
